feat(FPD-29): done create project page
