+++ src/interfaces/user/users.interface.ts
@@ -14,3 +14,34 @@ interface IGetUser {
 
 export type IGetUsers = IGetUser[];
 export type IUsers = IUser[];
+
+export interface DataType {
+  key: React.Key;
+  employeeId: string;
+  roles: string[];
+}
+
+export type ProjectType = {
+  name: string;
+  description: string;
+  date: string[];
+  language: string[];
+  technical: string[];
+  managerId: string;
+  employeeId: {
+    id: string;
+    roles: string[];
+  };
+};
+
+export type UserType = {
+  id: string;
+  userName: string;
+  email: string;
+  isManager: boolean;
+};
+
+export type SkillType = {
+  id: string;
+  name: string;
+};
